#1805 Implemented MDT import details parsing for paths, brushlines and icons. Fixed import of lines not showing up properly. Fixed other issues with MDT importing.
The round function contained some issue which rounded off the parsed values incorrectly and rounding them off to the nearest integer.

+++ app/Logic/MDT/Conversion.php
@@ -131,11 +131,11 @@ class Conversion
     /**
      * Rounds a number to the nearest two decimals.
      * @param $nr
-     * @return int
+     * @return float
      */
-    private static function round($nr): int
+    private static function round($nr): float
     {
-        return (int)($nr * 100) / 100;
+        return ((int)($nr * 100)) / 100;
     }
 
     /**
